requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
alabaster==0.7.12
astroid==1.6.6 ; python_version <= "3.0"  # via pylint-celery
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via pytest
babel==2.7.0
backports.functools-lru-cache==1.6.1  # via isort
certifi==2019.9.11
chardet==3.0.4
click-log==0.3.2          # via edx-lint
click==7.0                # via click-log, edx-lint
configparser==4.0.2       # via importlib-metadata
contextlib2==0.6.0.post1  # via importlib-metadata
coverage==4.5.4           # via pytest-cov
ddt==1.2.1
django-model-utils==3.2.0
django-nose==1.4.6
docutils==0.15.2
edx-lint==1.4.1
factory-boy==2.12.0
faker==2.0.4              # via factory-boy
filelock==3.0.12          # via tox
freezegun==0.3.12
funcsigs==1.0.2           # via mock, pytest
futures==3.3.0 ; python_version == "2.7"  # via isort
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23  # via pluggy, pytest, tox
ipaddress==1.0.23         # via faker
isort==4.3.21
jinja2==2.10.3
jsonfield==2.0.2
markupsafe==1.1.1
mock==3.0.5
more-itertools==5.0.0     # via pytest, zipp
nose==1.3.7
packaging==19.2
pathlib2==2.3.5           # via importlib-metadata, pytest, pytest-django
pluggy==0.13.1            # via pytest, tox
pockets==0.9.1
py==1.8.0                 # via pytest, tox
pycodestyle==2.5.0
pygments==2.4.2
pylint-celery==0.3        # via edx-lint
pylint-django==0.11.1 ; python_version < "3.0"  # via edx-lint
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==1.9.5 ; python_version < "3.0"  # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pyparsing==2.4.5
pytest-cov==2.8.1
pytest-django==3.7.0
pytest==4.6.6             # via pytest-cov, pytest-django
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scandir==1.10.0           # via pathlib2
six==1.13.0
snowballstemmer==2.0.0
sphinx-rtd-theme==0.4.3
sphinx==1.8.5
sphinxcontrib-napoleon==0.7
sphinxcontrib-websupport==1.1.2
text-unidecode==1.3       # via faker
toml==0.10.0              # via tox
tox==3.14.1
typing==3.7.4.1
urllib3==1.25.7
virtualenv==16.7.8        # via tox
wcwidth==0.1.7            # via pytest
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
